+++ src/components/Main/Main.tsx
@@ -3,6 +3,7 @@ import { Switch, Route } from 'react-router-dom';
 
 import Favorites from 'components/Favorites/Favorites';
 // Components
+import MediaPage from 'components/MediaPage/MediaPage';
 import Catalog from '../Catalog/Catalog';
 import Editor from '../Editor/Editor';
 // Styles
@@ -17,6 +18,7 @@ const Main: FC = () => {
         <Route exact path="/" component={Catalog} />
         <Route path="/editor" component={Editor} />
         <Route path="/favorites" component={Favorites} />
+        <Route path="/car/:carId" component={MediaPage} />
       </Switch>
     </main>
   );
